requirements.txt
python-telegram-bot
digitalguide
Pillow
PyYAML